chore: make transfer share logic rather than duplicating
