+++ frontend/components/receipts/transfer/basic-transfer-view.tsx
@@ -1,20 +1,18 @@
 import { AmountWithTooltip } from '../utils/string-amounts/amount-with-tooltip'
-import { VegaKey } from '../../keys/vega-key'
 import { ReceiptComponentProps } from '../receipts'
+import { BaseTransferView } from './base-transfer-view'
 
 export const locators = {
-  basicSection: 'basic-section'
+  basicTransferView: 'basic-transfer-view'
 }
 
 export const BasicTransferView = ({ transaction }: ReceiptComponentProps) => {
   const { asset, amount } = transaction.transfer
   return (
-    <div data-testid={locators.basicSection}>
+    <BaseTransferView transaction={transaction}>
       <div className="text-xl text-white">
         <AmountWithTooltip assetId={asset} amount={amount} />
       </div>
-      <h1 className="text-vega-dark-300 mt-4">To</h1>
-      <VegaKey publicKey={transaction.transfer.to} name="Receiving Key" />
-    </div>
+    </BaseTransferView>
   )
 }
