+++ frontend/components/receipts/transfer/enriched-transfer-view.tsx
@@ -1,9 +1,8 @@
 import { formatNumber, toBigNum } from '@vegaprotocol/utils'
 import { PriceWithSymbol } from '../utils/string-amounts/price-with-symbol'
-import { VegaKey } from '../../keys/vega-key'
-import { useWalletStore } from '../../../stores/wallets'
 import { ReceiptComponentProps } from '../receipts'
 import { useAssetsStore } from '../../../stores/assets-store'
+import { BaseTransferView } from './base-transfer-view'
 
 export const locators = {
   enrichedSection: 'enriched-section'
@@ -14,21 +13,18 @@ export const EnrichedTransferView = ({ transaction }: ReceiptComponentProps) => 
     getAssetById: state.getAssetById
   }))
   const assetInfo = getAssetById(transaction.transfer.asset)
-  const { getKeyById } = useWalletStore((state) => ({ getKeyById: state.getKeyById }))
   const { amount } = transaction.transfer
   const decimals = Number(assetInfo?.details?.decimals)
-  const price = amount && decimals ? formatNumber(toBigNum(amount, decimals), decimals) : undefined
+  const formattedAmount = amount && decimals ? formatNumber(toBigNum(amount, decimals), decimals) : undefined
   const symbol = assetInfo?.details?.symbol
-  const keyInfo = getKeyById(transaction.transfer.to)
-  const isOwnKey = keyInfo?.publicKey === transaction.transfer.to
 
   return (
-    <div data-testid={locators.enrichedSection}>
-      <div className="text-xl text-white">
-        {price && symbol ? <PriceWithSymbol price={price} symbol={symbol} /> : null}
+    <BaseTransferView transaction={transaction}>
+      <div data-testid={locators.enrichedSection}>
+        <div className="text-xl text-white">
+          {formattedAmount && symbol ? <PriceWithSymbol price={formattedAmount} symbol={symbol} /> : null}
+        </div>
       </div>
-      <h1 className="text-vega-dark-300 mt-4">To</h1>
-      <VegaKey publicKey={transaction.transfer.to} name={isOwnKey ? `${keyInfo?.name} (own key)` : 'External key'} />
-    </div>
+    </BaseTransferView>
   )
 }
